rename notifications to cart
