+++ app/app/src/main/java/com/ar/enbaldeapp/ui/cart/CartFragment.java
@@ -1,4 +1,4 @@
-package com.ar.enbaldeapp.ui.notifications;
+package com.ar.enbaldeapp.ui.cart;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -10,22 +10,22 @@ import androidx.annotation.NonNull;
 import androidx.fragment.app.Fragment;
 import androidx.lifecycle.ViewModelProvider;
 
-import com.ar.enbaldeapp.databinding.FragmentNotificationsBinding;
+import com.ar.enbaldeapp.databinding.FragmentCartBinding;
 
-public class NotificationsFragment extends Fragment {
+public class CartFragment extends Fragment {
 
-    private FragmentNotificationsBinding binding;
+    private FragmentCartBinding binding;
 
     public View onCreateView(@NonNull LayoutInflater inflater,
                              ViewGroup container, Bundle savedInstanceState) {
-        NotificationsViewModel notificationsViewModel =
-                new ViewModelProvider(this).get(NotificationsViewModel.class);
+        CartViewModel cartViewModel =
+                new ViewModelProvider(this).get(CartViewModel.class);
 
-        binding = FragmentNotificationsBinding.inflate(inflater, container, false);
+        binding = FragmentCartBinding.inflate(inflater, container, false);
         View root = binding.getRoot();
 
-        final TextView textView = binding.textNotifications;
-        notificationsViewModel.getText().observe(getViewLifecycleOwner(), textView::setText);
+        final TextView textView = binding.textCart;
+        cartViewModel.getText().observe(getViewLifecycleOwner(), textView::setText);
         return root;
     }
