+++ app/app/src/main/java/com/ar/enbaldeapp/ui/cart/CartViewModel.java
@@ -1,16 +1,16 @@
-package com.ar.enbaldeapp.ui.notifications;
+package com.ar.enbaldeapp.ui.cart;
 
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
 
-public class NotificationsViewModel extends ViewModel {
+public class CartViewModel extends ViewModel {
 
     private final MutableLiveData<String> mText;
 
-    public NotificationsViewModel() {
+    public CartViewModel() {
         mText = new MutableLiveData<>();
-        mText.setValue("This is notifications fragment");
+        mText.setValue("This is cart fragment");
     }
 
     public LiveData<String> getText() {
