+++ app/app/src/main/res/menu/bottom_nav_menu.xml
@@ -12,9 +12,9 @@
         android:title="@string/title_catalogue" />
 
     <item
-        android:id="@+id/navigation_notifications"
-        android:icon="@drawable/ic_notifications_black_24dp"
-        android:title="@string/title_notifications" />
+        android:id="@+id/navigation_cart"
+        android:icon="@drawable/baseline_shopping_cart_24"
+        android:title="@string/title_cart" />
 
     <item
         android:id="@+id/navigation_profile"
