+++ app/app/src/main/res/values/strings.xml
@@ -2,7 +2,7 @@
     <string name="app_name">Enbalde</string>
     <string name="title_home">Home</string>
     <string name="title_catalogue">Catalogue</string>
-    <string name="title_notifications">Notifications</string>
+    <string name="title_cart">Cart</string>
     <!-- TODO: Remove or change this placeholder text -->
     <string name="hello_blank_fragment">Hello blank fragment</string>
 </resources>
